# src/main/webui/openapi/openapi.yaml
...
openapi: 3.1.0
components:
  schemas:
...
    CreateTeamRequest:
      type: object
      properties:
        name:
          type: string
        brandColor:
          type: string
...
    Instant:
      type: string
      format: date-time
      examples:
      - 2022-03-10T16:15:50Z
...
    MediaType:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        url:
          type: string
        mimeType:
          type: string
...
    TeamType:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        name:
          type: string
        brandColor:
          type: string
        coverMedia:
          $ref: "#/components/schemas/MediaType"
        revealTributes:
          type: boolean
        revealAt:
          $ref: "#/components/schemas/Instant"
        tributesRevealed:
          type: boolean
        yearbookTitle:
          type: string
        yearbookSubtitle:
          type: string
        yearbookDedication:
          type: string
        yearbookTheme:
          $ref: "#/components/schemas/YearbookTheme"
        yearbookShowMembers:
          type: boolean
        yearbookShowTributes:
          type: boolean
        yearbookShowCharacteristics:
          type: boolean
        yearbookShowMemories:
          type: boolean
        yearbookShowAwards:
          type: boolean
...
    UUID:
      type: string
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
...
    YearbookTheme:
      type: string
      enum:
      - CLASSIC
      - MODERN
      - SCRAPBOOK
      - MINIMAL
...
  securitySchemes:
    SecurityScheme:
      type: openIdConnect
      description: Authentication
...
paths:
...
  /api/teams:
    get:
      summary: List teams for the current user
      tags:
      - Teams
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamType"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - user
    post:
      summary: Create a team
      tags:
      - Teams
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamType"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      security:
      - SecurityScheme:
        - user
...
info:
  title: Yaadbuzz REST API
  version: 1.0.0
  description: "Auth, media, and yearbook domain REST APIs for Yaadbuzz."